backend/requirements.txt
# Django and REST Framework
Django==5.2.6
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1

# Database
dj-database-url==3.0.1

# CORS handling
django-cors-headers==4.8.0

# Environment variables
python-decouple==3.8

# Firebase
firebase-admin==6.5.0

# Static files
whitenoise==6.10.0

# Production server
gunicorn==21.2.0

# Additional utilities
Pillow==11.0.0
requests==2.32.5

# Development tools (optional)
django-debug-toolbar==4.4.6

# Admin interface
django-jazzmin